add generic text input styling
text inputs are now styled at every occurence in the app,
style input by default, and provide a .mx_textinput class
if buttons need to appear inside the input styling.

Before this was partially done in _common.scss, but as it's
highly theme-dependent, makes more sense to do it in the theme.