Clarify readme instructions for developers
Setting up the SDK for development, I noticed that 3 tests failed on
develop branch. Linking the JS SDK made the tests pass. It looks like
the assumption is that developers link against the develop branch
of JS SDK to develop the React SDK.

Clarify this in the readme for completeness sake.

Signed-off-by: Jason Robinson <jasonr@matrix.org>