Key Backup: Don't fail if no keys
Only fail if there were any keys in the backup (which does mean
that the backup will always succeed if there aren't any keys which
could also be misleading, but is probably not as bad and can probably
be fixed with Trust on Decrypt).

Also cheekily fix the error message so it talks about passphrases
if you used a passphrase and recovery keys if you used a recovery key.