Bring back the lowered opacity for offline/unavailable avatars.
At some point during the riot redesign, the bit of css that lowered the opacity for offline/unavailable contacts
was removed. This makes it impossible to discern presence for a contact unless you hover your cursor over their
avatar.

It's very handy to be able to see presence at a glance without any hovering, so this PR reintroduces the lowered opacity.
I've also slightly decreased the opacity level from 0.66 to 0.5 to make it slightly more noticable.