Send Field label pointer events to input
When the `label` element is displayed on top of the input (`label` is set and
there is no `placeholder`), it would block clicks from reaching the input. This
allows them to get through, but then also restores `label`'s events once it
moves out of the way.

Fixes https://github.com/vector-im/riot-web/issues/8469